modif for mesonet

+++ retinotopy.py
@@ -198,8 +198,8 @@ class Params:
     
     datetag: str = datetag # Set the date of the result's file
     loader: str = f'{DATAROOT}/Imagenet_urls_ILSVRC_2016.json' # File containing Imagenet's labels
-    annotations_animal: str = f'{DATAROOT}/Animal10k_annotations.json' # File containing Animak10k's labels
-    annotations_val: str = f'{DATAROOT}/LOC_val_solution_with_sizes.csv' # File containing Imagenets's labels
+    # annotations_animal: str = f'{DATAROOT}/Animal10k_annotations.json' # File containing Animak10k's labels
+    # annotations_val: str = f'{DATAROOT}/LOC_val_solution_with_sizes.csv' # File containing Imagenets's labels
 
     # root: str = f'{DATAROOT}/Imagenet_{data_set_type}' # Directory containing images to perform the training
     folders: list = field(default_factory=lambda: ['val', 'train']) # Set the training and validation folders relative to the root
@@ -242,9 +242,10 @@ class Params:
 args = Params()
 
 # keep a human readable copy of the parameters
-json_fname = os.path.join(data_cache, strftime("%Y-%m-%d", gmtime()) + '_config_args.json')
-with open(json_fname, 'wt') as f:
-    json.dump(vars(args), f, indent=4)
+json_fname = os.path.join(data_cache, datetag + '_config_args.json')
+if not(os.path.isfile(json_fname)):
+    with open(json_fname, 'wt') as f:
+        json.dump(vars(args), f, indent=4)
 #############################################################
 
 
@@ -284,6 +285,10 @@ for i_img, img_id in enumerate(Imagenet_urls_ILSVRC_2016):
 
 #---------------Get annotations from other the data set (Animal10k, ...)---------------
 def get_annotation(type):
+
+    annotations_animal: str = f'{DATAROOT}/Animal10k_annotations.json' # File containing Animak10k's labels
+    annotations_val: str = f'{DATAROOT}/LOC_val_solution_with_sizes.csv' # File containing Imagenets's labels
+
     if 'csv' in type:
         with open(args.annotations_val, 'r') as csv_file:
             return pd.read_csv(csv_file)
